// src/types/item.ts
export interface ItemResponse {
  id: string;
  category_id: string;
  category_name?: string;
  name: string;
  code: string; // read-only, auto-generated by BE
  variant: string | null;
  description: string | null;
  unit: string | null;
  created_at: string;
  updated_at: string;
}

export interface ItemListResponse {
  success: boolean;
  data: ItemResponse[];
  paging: {
    hasNext: boolean;
    currentPage: number;
    totalPages: number;
    totalData: number;
    limit: number;
  };
}